change: refactor to 7 modules
