+++ src/modules/2-nwd.module.js
@@ -1,0 +1,3 @@
+export default {
+
+}
