+++ src/modules/3-files.module.js
@@ -1,0 +1,3 @@
+export default {
+
+}
