+++ src/modules/6-zip.module.js
@@ -1,0 +1,3 @@
+export default {
+  
+}
