--- src/utils/index.js
@@ -1,9 +1,0 @@
-import { exitProgram } from './exitProgram.util.js';
-import { getUsername } from './getUsername.util.js';
-import { printHello } from './printHello.util.js';
-
-export {
-  exitProgram,
-  getUsername,
-  printHello,
-}
